requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==2.3.6
flask_sqlalchemy
flask_migrate
flask_login
psycopg2
